--- autogpt_platform/autogpt_libs/autogpt_libs/supabase_integration_credentials_store/__init__.py
@@ -1,9 +1,0 @@
-from .store import SupabaseIntegrationCredentialsStore
-from .types import APIKeyCredentials, Credentials, OAuth2Credentials
-
-__all__ = [
-    "SupabaseIntegrationCredentialsStore",
-    "Credentials",
-    "APIKeyCredentials",
-    "OAuth2Credentials",
-]
