+++ autogpt_platform/backend/backend/blocks/ai_shortform_video_block.py
@@ -3,11 +3,15 @@ import time
 from enum import Enum
 from typing import Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 from backend.util.request import requests
 
 TEST_CREDENTIALS = APIKeyCredentials(
