+++ autogpt_platform/backend/backend/blocks/hubspot/_auth.py
@@ -1,9 +1,8 @@
 from typing import Literal
 
-from autogpt_libs.supabase_integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
-from backend.data.model import CredentialsField, CredentialsMetaInput
+from backend.data.model import APIKeyCredentials, CredentialsField, CredentialsMetaInput
 
 HubSpotCredentials = APIKeyCredentials
 HubSpotCredentialsInput = CredentialsMetaInput[
