+++ autogpt_platform/backend/backend/integrations/creds_manager.py
@@ -2,14 +2,12 @@ import logging
 from contextlib import contextmanager
 from datetime import datetime
 
-from autogpt_libs.supabase_integration_credentials_store import (
-    Credentials,
-    SupabaseIntegrationCredentialsStore,
-)
 from autogpt_libs.utils.synchronize import RedisKeyedMutex
 from redis.lock import Lock as RedisLock
 
 from backend.data import redis
+from backend.data.model import Credentials
+from backend.integrations.credentials_store import IntegrationCredentialsStore
 from backend.integrations.oauth import HANDLERS_BY_NAME, BaseOAuthHandler
 from backend.util.exceptions import MissingConfigError
 from backend.util.settings import Settings
@@ -53,7 +51,7 @@ class IntegrationCredentialsManager:
     def __init__(self):
         redis_conn = redis.get_redis()
         self._locks = RedisKeyedMutex(redis_conn)
-        self.store = SupabaseIntegrationCredentialsStore(redis=redis_conn)
+        self.store = IntegrationCredentialsStore()
 
     def create(self, user_id: str, credentials: Credentials) -> None:
         return self.store.add_creds(user_id, credentials)
