+++ autogpt_platform/frontend/src/lib/autogpt-server-api/types.ts
@@ -289,7 +289,7 @@ export type CredentialsMetaInput = {
   provider: string;
 };
 
-/* Mirror of autogpt_libs/supabase_integration_credentials_store/types.py:_BaseCredentials */
+/* Mirror of backend/backend/data/model.py:_BaseCredentials */
 type BaseCredentials = {
   id: string;
   type: CredentialsType;
@@ -297,7 +297,7 @@ type BaseCredentials = {
   provider: CredentialsProviderName;
 };
 
-/* Mirror of autogpt_libs/supabase_integration_credentials_store/types.py:OAuth2Credentials */
+/* Mirror of backend/backend/data/model.py:OAuth2Credentials */
 export type OAuth2Credentials = BaseCredentials & {
   type: "oauth2";
   scopes: string[];
@@ -309,7 +309,7 @@ export type OAuth2Credentials = BaseCredentials & {
   metadata: Record<string, any>;
 };
 
-/* Mirror of autogpt_libs/supabase_integration_credentials_store/types.py:APIKeyCredentials */
+/* Mirror of backend/backend/data/model.py:APIKeyCredentials */
 export type APIKeyCredentials = BaseCredentials & {
   type: "api_key";
   title: string;
